converted scale to double from int and minor fixes related to bomb positioning

+++ src/bomberman/entity/Entity.java
@@ -28,5 +28,5 @@ public interface Entity {
      * */
     int getLayer();
 
-    int getScale();
+    double getScale();
 }
